# boutique_artisans_2/boutique_artisans/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "API for managing shops",
        "version": "1.0.0"
    },
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of all products",
                "description": "Returns list of products",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product details by ID",
                "description": "Returns details of a specific product by ID.",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "9x508988-2d3g-44f6-ac72-0f93eac22rrd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "text/html": {}
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/shops": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Get list of all shops",
                "description": "Returns a list of all shops.",
                "operationId": "getShops",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "text/html": {}
                        }
                    }
                }
            }
        }
    }
}